총 게이지 완료
완료 작업
1. 총게이지

다음 작업
1. 스킬 관리 시스템

+++ Assets/02Scripts/InGame/Vation/Item/DataSO/Item/Equipment/Weapon/WeaponItemDataSO.cs
@@ -31,7 +31,7 @@ namespace alpha
 #endif
         [Header("[ WeaponItemData ]"), Space(10)]
         public int Damage;
-        public int AttackCost;
+        public float AttackCost;
         public AudioClip AudioClip;
     }
 }
